members/requirements.txt
Django==4.2
psycopg==3.1.18
pycryptodomex==3.20.0
requests==2.31.0
python-dateutil==2.9.0.post0
